+++ app/controllers/guilds_controller.rb
@@ -1,4 +1,5 @@
 class GuildsController < ApplicationController
+  before_action :find_guild_invites_for_user, only: %i[index]
   before_action :find_guild, only: %i[show management]
   before_action :find_user_characters, only: %i[show]
 
@@ -11,6 +12,10 @@ class GuildsController < ApplicationController
   end
 
   private
+
+  def find_guild_invites_for_user
+    @guild_invites = Current.user.guild_invites
+  end
 
   def find_guild
     @guild = Guild.find_by(slug: params[:id])