+++ app/models/guild_invite.rb
@@ -1,5 +1,7 @@
 # Represents invite to guild and requests for guild membership
 class GuildInvite < ApplicationRecord
+  enum status: { send: 0, declined: 1 }, _prefix: :status
+
   belongs_to :guild
   belongs_to :character