+++ app/policies/guild_invite_policy.rb
@@ -14,4 +14,13 @@ class GuildInvitePolicy < ApplicationPolicy
   def destroy?
     new?
   end
+
+  def approve?
+    return character.user_id == user.id && character.guild_id.nil? if record == 'true'
+    user.any_role?(guild.id, 'gm', 'rl')
+  end
+
+  def decline?
+    approve?
+  end
 end