+++ config/routes.rb
@@ -64,7 +64,10 @@ Rails.application.routes.draw do
       get :decline, on: :member
     end
     resources :deliveries, only: %i[new destroy]
-    resources :guild_invites, only: %i[new destroy]
+    resources :guild_invites, only: %i[new destroy] do
+      get :approve, on: :member
+      get :decline, on: :member
+    end
 
     root to: 'welcome#index'
   end