+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - page for guild invites
 - api endpoint for creating guild invites
 - deleting guild invites
+- approving/declining guild invites by users
 
 ## [0.6.4] - 2019-11-06
 ### Added